requirements.txt
langchain-core>=0.1.27
langchain-community>=0.0.24
langgraph>=0.0.24
starlette>=0.36.3
uvicorn[standard]>=0.27.1
fastapi[standard]>=0.109.2
tavily-python>=0.3.1
websockets>=12.0
python-dotenv>=1.0.1
openai>=1.12.0

